fix(directus-node): separate files to allow imports in middleware
